refactor(pi-exa): split monolithic extension into modular structure
- Split 850-line index.ts into focused modules:
  - config.ts: Config loading, auth resolution, tool enablement
  - constants.ts: DEFAULT_MAX_CHARACTERS, DEFAULT_NUM_RESULTS
  - formatters.ts: formatSearchResults, formatCrawlResults
  - types.ts: Shared types and helpers
  - tools/: Individual tool implementations (web-search, web-fetch, web-search-advanced)
- Maintain backward-compatible exports for existing tests
- All 72 tests pass, lint and typecheck clean
