+++ packages/pi-exa/extensions/constants.ts
@@ -1,0 +1,6 @@
+/**
+ * Exa Extension Constants
+ */
+
+export const DEFAULT_MAX_CHARACTERS = 3000;
+export const DEFAULT_NUM_RESULTS = 5;
