+++ packages/pi-exa/extensions/tools/index.ts
@@ -1,0 +1,12 @@
+/**
+ * Tools barrel export
+ */
+
+export {
+  webFetchParams,
+  webSearchAdvancedParams,
+  webSearchParams,
+} from "./definitions.js";
+export { createWebFetchTool } from "./web-fetch.js";
+export { createWebSearchTool } from "./web-search.js";
+export { createWebSearchAdvancedTool } from "./web-search-advanced.js";
